Split protocols module into frame + codec modules
